backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.3

python-multipart==0.0.9
httpx==0.27.0
